+++ internal/config/config_test.go
@@ -19,7 +19,7 @@ func configFileReader(content string) io.Reader {
 
 func TestLoadClearPrevConfig(t *testing.T) {
 	Config = config{SocketPath: "/tmp"}
-	err := Load(nil)
+	err := Load(&bytes.Buffer{})
 	assert.NoError(t, err)
 
 	assert.Empty(t, Config.SocketPath)
@@ -144,7 +144,7 @@ func TestLoadOnlyEnvironment(t *testing.T) {
 	os.Setenv("GITALY_LISTEN_ADDR", ":8081")
 	os.Setenv("GITALY_PROMETHEUS_LISTEN_ADDR", ":9237")
 
-	err := Load(nil)
+	err := Load(&bytes.Buffer{})
 	assert.NoError(t, err)
 
 	assert.Equal(t, ":9237", Config.PrometheusListenAddr)
@@ -297,6 +297,7 @@ func TestValidateShellPath(t *testing.T) {
 
 	tmpDir, err := ioutil.TempDir("", "gitaly-tests-")
 	require.NoError(t, err)
+	require.NoError(t, os.MkdirAll(path.Join(tmpDir, "bin"), 0755))
 	tmpFile := path.Join(tmpDir, "my-file")
 	defer os.RemoveAll(tmpDir)
 	fp, err := os.Create(tmpFile)
@@ -311,7 +312,7 @@ func TestValidateShellPath(t *testing.T) {
 		{
 			desc:      "When no Shell Path set",
 			path:      "",
-			shouldErr: false,
+			shouldErr: true,
 		},
 		{
 			desc:      "When Shell Path set to non-existing path",
@@ -340,5 +341,71 @@ func TestValidateShellPath(t *testing.T) {
 			assert.NoError(t, err)
 		}
 
+	}
+}
+
+func TestValidateRuby(t *testing.T) {
+	defer func(oldRuby Ruby) {
+		Config.Ruby = oldRuby
+	}(Config.Ruby)
+
+	tmpDir, err := ioutil.TempDir("", "gitaly-test")
+	require.NoError(t, err)
+	defer os.RemoveAll(tmpDir)
+
+	tmpFile := path.Join(tmpDir, "file")
+	require.NoError(t, ioutil.WriteFile(tmpFile, nil, 0644))
+
+	testCases := []struct {
+		dir  string
+		ok   bool
+		desc string
+	}{
+		{dir: "", desc: "empty"},
+		{dir: "/does/not/exist", desc: "does not exist"},
+		{dir: tmpFile, desc: "exists but is not a directory"},
+		{dir: tmpDir, ok: true, desc: "ok"},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.desc, func(t *testing.T) {
+			Config.Ruby = Ruby{Dir: tc.dir}
+
+			err := validateRuby()
+			if tc.ok {
+				require.NoError(t, err)
+			} else {
+				require.Error(t, err)
+			}
+		})
+	}
+}
+
+func TestValidateListeners(t *testing.T) {
+	defer func(cfg config) {
+		Config = cfg
+	}(Config)
+
+	testCases := []struct {
+		desc string
+		config
+		ok bool
+	}{
+		{desc: "empty"},
+		{desc: "socket only", config: config{SocketPath: "/foo/bar"}, ok: true},
+		{desc: "tcp only", config: config{ListenAddr: "a.b.c.d:1234"}, ok: true},
+		{desc: "both socket and tcp", config: config{SocketPath: "/foo/bar", ListenAddr: "a.b.c.d:1234"}, ok: true},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.desc, func(t *testing.T) {
+			Config = tc.config
+			err := validateListeners()
+			if tc.ok {
+				require.NoError(t, err)
+			} else {
+				require.Error(t, err)
+			}
+		})
 	}
 }