+++ internal/service/register.go
@@ -8,8 +8,8 @@ import (
 	"gitlab.com/gitlab-org/gitaly/internal/service/diff"
 	"gitlab.com/gitlab-org/gitaly/internal/service/namespace"
 	"gitlab.com/gitlab-org/gitaly/internal/service/notifications"
+	"gitlab.com/gitlab-org/gitaly/internal/service/operations"
 	"gitlab.com/gitlab-org/gitaly/internal/service/ref"
-	"gitlab.com/gitlab-org/gitaly/internal/service/renameadapter"
 	"gitlab.com/gitlab-org/gitaly/internal/service/repository"
 	"gitlab.com/gitlab-org/gitaly/internal/service/smarthttp"
 	"gitlab.com/gitlab-org/gitaly/internal/service/ssh"
@@ -22,40 +22,16 @@ import (
 // RegisterAll will register all the known grpc services with
 // the specified grpc service instance
 func RegisterAll(grpcServer *grpc.Server, rubyServer *rubyserver.Server) {
-	notificationsService := notifications.NewServer()
-	pb.RegisterNotificationServiceServer(grpcServer, notificationsService)
-
-	refService := ref.NewServer(rubyServer)
-	pb.RegisterRefServiceServer(grpcServer, refService)
-
-	smartHTTPService := smarthttp.NewServer()
-	pb.RegisterSmartHTTPServiceServer(grpcServer, smartHTTPService)
-
-	diffService := diff.NewServer(rubyServer)
-	pb.RegisterDiffServiceServer(grpcServer, diffService)
-
-	commitService := commit.NewServer(rubyServer)
-	pb.RegisterCommitServiceServer(grpcServer, commitService)
-
-	sshService := ssh.NewServer()
-	pb.RegisterSSHServiceServer(grpcServer, sshService)
-
-	blobService := blob.NewServer()
-	pb.RegisterBlobServiceServer(grpcServer, blobService)
-
-	repositoryService := repository.NewServer()
-	pb.RegisterRepositoryServiceServer(grpcServer, repositoryService)
-
-	namespaceService := namespace.NewServer()
-	pb.RegisterNamespaceServiceServer(grpcServer, namespaceService)
-
-	// Deprecated Services
-	pb.RegisterNotificationsServer(grpcServer, renameadapter.NewNotificationAdapter(notificationsService))
-	pb.RegisterRefServer(grpcServer, renameadapter.NewRefAdapter(refService))
-	pb.RegisterSmartHTTPServer(grpcServer, renameadapter.NewSmartHTTPAdapter(smartHTTPService))
-	pb.RegisterDiffServer(grpcServer, renameadapter.NewDiffAdapter(diffService))
-	pb.RegisterCommitServer(grpcServer, renameadapter.NewCommitAdapter(commitService))
-	pb.RegisterSSHServer(grpcServer, renameadapter.NewSSHAdapter(sshService))
+	pb.RegisterBlobServiceServer(grpcServer, blob.NewServer())
+	pb.RegisterCommitServiceServer(grpcServer, commit.NewServer(rubyServer))
+	pb.RegisterDiffServiceServer(grpcServer, diff.NewServer(rubyServer))
+	pb.RegisterNamespaceServiceServer(grpcServer, namespace.NewServer())
+	pb.RegisterNotificationServiceServer(grpcServer, notifications.NewServer())
+	pb.RegisterOperationServiceServer(grpcServer, operations.NewServer(rubyServer))
+	pb.RegisterRefServiceServer(grpcServer, ref.NewServer(rubyServer))
+	pb.RegisterRepositoryServiceServer(grpcServer, repository.NewServer(rubyServer))
+	pb.RegisterSSHServiceServer(grpcServer, ssh.NewServer())
+	pb.RegisterSmartHTTPServiceServer(grpcServer, smarthttp.NewServer())
 
 	healthpb.RegisterHealthServer(grpcServer, health.NewServer())
 }