+++ internal/service/repository/server.go
@@ -1,12 +1,16 @@
 package repository
 
 import (
+	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
+
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
 )
 
-type server struct{}
+type server struct {
+	*rubyserver.Server
+}
 
 // NewServer creates a new instance of a gRPC repo server
-func NewServer() pb.RepositoryServiceServer {
-	return &server{}
+func NewServer(rs *rubyserver.Server) pb.RepositoryServiceServer {
+	return &server{rs}
 }